dev merge

+++ back/src/services/storyPostService.js
@@ -33,76 +33,6 @@ class storyPostService {
     return createdNewStoryPost;
   }
 
-  // static async updateStory({ storyId, toUpdate }) {
-  //   let story = await storyPostModel.findOneByStoryId({ storyId });
-
-  //   if (!story) {
-  //     throw new Error('해당 스토리를 찾을 수 없습니다. 다시 한번 확인해주세요');
-  //   }
-
-  //   if (toUpdate.title) {
-  //     const fieldToUpdate = 'title';
-  //     const newValue = toUpdate.title;
-  //     story = await storyPostModel.updateStory({
-  //       storyId,
-  //       fieldToUpdate,
-  //       newValue,
-  //     });
-  //   }
-
-  //   if (toUpdate.content) {
-  //     const fieldToUpdate = 'content';
-  //     const newValue = toUpdate.content;
-  //     story = await storyPostModel.updateStory({
-  //       storyId,
-  //       fieldToUpdate,
-  //       newValue,
-  //     });
-  //   }
-
-  //   if (toUpdate.isPublic) {
-  //     const fieldToUpdate = 'isPublic';
-  //     const newValue = toUpdate.isPublic;
-  //     story = await storyPostModel.updateStory({
-  //       storyId,
-  //       fieldToUpdate,
-  //       newValue,
-  //     });
-  //   }
-
-  //   if (toUpdate.thumbnail) {
-  //     // create Image
-  //     const fieldToUpdate = 'thumbnail';
-  //     const newValue = toUpdate.thumbnail;
-  //     story = await storyPostModel.updateStory({
-  //       storyId,
-  //       fieldToUpdate,
-  //       newValue,
-  //     });
-  //   }
-
-  //   if (toUpdate.mood) {
-  //     const fieldToUpdate = 'mood';
-  //     const newValue = toUpdate.mood;
-  //     story = await storyPostModel.updateStory({
-  //       storyId,
-  //       fieldToUpdate,
-  //       newValue,
-  //     });
-  //   }
-
-  //   if (toUpdate.music) {
-  //     const fieldToUpdate = 'music';
-  //     const newValue = toUpdate.music;
-  //     story = await storyPostModel.updateStory({
-  //       storyId,
-  //       fieldToUpdate,
-  //       newValue,
-  //     });
-  //   }
-  //   return story;
-  // }
-
   static async deleteStory({ storyId }) {
     let isDeleted = await storyPostModel.deleteOneByStoryId({ storyId });
     if (!isDeleted) {
@@ -120,7 +50,7 @@ class storyPostService {
     }
 
     const storyInfo = {
-      ...story._doc, // document를 자바스크립트 객체로 변환하기 위해 사용되는 속성
+      ...story,
       // commentCount: allComments.length,
       commentList: allComments,
     };
@@ -150,6 +80,32 @@ class storyPostService {
     return { stories, totalPage, count }; // 해당 페이지에 해당하는 스토리들, 총 페이지 수, 스토리 총 수
   }
 
+  static async readMyPosts(limit, page, userId) {
+    const skip = (page - 1) * limit; // 해당 페이지에서 스킵할 스토리 수
+
+    const { stories, count } = await storyPostModel.findMyAndCountAll(
+      skip,
+      limit,
+      userId,
+    );
+    const totalPage = Math.ceil(count / limit);
+    return { stories, totalPage, count }; // 해당 페이지에 해당하는 스토리들, 총 페이지 수, 스토리 총 수
+  }
+
+  static async readMySearchQueryPosts(limit, page, userId, searchQuery) {
+    const skip = (page - 1) * limit; // 해당 페이지에서 스킵할 스토리 수
+
+    const { stories, count } =
+      await storyPostModel.findMySearchQueryAndCountAll(
+        skip,
+        limit,
+        userId,
+        searchQuery,
+      );
+    const totalPage = Math.ceil(count / limit);
+    return { stories, totalPage, count }; // 해당 페이지에 해당하는 스토리들, 총 페이지 수, 스토리 총 수
+  }
+
   static async populateStoryPost(info, path) {
     const field = { path: path };
     const result = storyPostModel.populateStoryPost(info, field);
@@ -166,16 +122,19 @@ class storyPostService {
     }
   }
 
-  static async deleteUploadImage({ storyId }) {
+  static async deleteUploadedImage({ storyId }) {
     const story = await storyPostModel.findOneByStoryId({ storyId });
     const uploadImage = await imageModel.findOneByImageId({
       imageId: story.thumbnail,
     });
-    const uploadImagePath = uploadImage.path;
-    if (fs.existsSync(uploadImagePath)) {
-      fs.unlinkSync(uploadImagePath);
+    if (uploadImage) {
+      const uploadImagePath = uploadImage.path;
+      if (fs.existsSync(uploadImagePath)) {
+        fs.unlinkSync(uploadImagePath);
+      } else {
+        console.log('File does not exist, so not deleting.');
+      }
     } else {
-      console.log('File does not exist, so not deleting.');
     }
   }
